raft: set threadpool timeout to be longer than heartbeat interval
Looking at a cluster where the heartbeat interval has been set to 1.5
seconds, I see that there are a lot of threads starting and stopping.
The issue seems to be that the Raft threadpool (on which requests are
scheduled) has a default 500ms timeout, but the requests are less
frequent. Thus, every request is starting and stopping a thread.

Change-Id: Ic1b11d3d9230af8fe8cc3490fed51738ae7e81a1
Reviewed-on: http://gerrit.cloudera.org:8080/4635
Tested-by: Kudu Jenkins
Reviewed-by: David Ribeiro Alves <dralves@apache.org>