[flaky tests] Fix ClientTest.TestFailedDnsResolution
In this test we're testing that we get dns resolution error back when
a dns resolution error fails. However there is a narrow window in which
the returned error might be a master GetTabletLocations() RPC instead,
because of KUDU-1466.

This changes the test to retry until it gets the appropriate error.

Change-Id: I9fa3ceb7215d60cd23efabf729731254524e1625
Reviewed-on: http://gerrit.cloudera.org:8080/4643
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Tested-by: David Ribeiro Alves <dralves@apache.org>