[util] added Subprocess::GetExitStatus()
The Subprocess::GetExitStatus() method allows to retrive exit status
and exit information in printable form.  That facilitates reporting
proper exit status for a sub-processes if using Subprocess::Wait().

The motivation for this change was to free the users of the Subprocess
interface of calling WIFEXITED/WEXITSTATUS and friends, duplicating
that code over and over.  In some cases, callers did not use
WEXITSTATUS or WTERMSIG and mistakenly interpreted wait status
as sub-process exit status.  In short, most users of the Subprocess
interface need exit status of a sub-process, not wait status.

Also, addressed the race when a signal arrives to the child process
before it calls execve().  Todd suggested to use the trick with
a pipe and O_CLOEXEC option.

Change-Id: Ic2b16e2a2a53a01982f816b9ee41cc61fd93d4bf
Reviewed-on: http://gerrit.cloudera.org:8080/4648
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Alexey Serbin <aserbin@cloudera.com>