[java client] Cleanup AsyncKuduClient's unused caches
Originally, asynchbase came with a few caches such as server-connection->region and
region->server-connection. Via dozens of patches, we've reduced their usefulness to 0.
This patch simple finishes the job and removes them.

FWIW client2tablets was always a source of pain, and the caching logic is now a lot
easier to manage, and potentially refactor!

Change-Id: I62802c34c618c83a4ff69d79825387cbe4ab51a8
Reviewed-on: http://gerrit.cloudera.org:8080/4705
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>