TODO cleanup in TransactionManager
This removes a couple of TODOs that were now wrong because we now
have exactly-once retry semantics.

The first was about not having to perform the actual waiting for commit
wait on followers. With the new retry semantics this could easily mean
that upon retrying a request on a follower a client would observe the
result of the operation before it safe to do so.

The second was about not peforming the waiting for commit wait if the
RPCs deadline was expired. A client can now have a total deadline that
is higher than a specific RPC's deadline and has exactly once semantics
meaning even if the specific RPC's deadline is expired the client might
still retry within the context of the global deadline. In that case
we wouldn't want to have the client observe the operation before it
was supposed to.

Change-Id: Ib41a5a65dee29d26f0e7c72b4f2d30f1267c8cf2
Reviewed-on: http://gerrit.cloudera.org:8080/4826
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>