consensus: remove bits of dead code
We had some old code path in which PeerMessageQueue::AppendOperations()
would previously return ServiceUnavailable indicating some kind of
memory limit. However, I traced through the code and could only find one
case in which it would return ServiceUnavailable: if the log itself is
shut down.

We expect that the lifecycle of the log is controlled such that this
will never happen. So, the code path that tried to handle this is now
just a CHECK, and I was able to remove some code that was only called in
this (dead) case.

Additionally removes some other unused typedefs and methods in
ReplicaState.

Change-Id: I8dc0a129f4f136256083a31e4b4b27e6a673dbee
Reviewed-on: http://gerrit.cloudera.org:8080/4712
Tested-by: Kudu Jenkins
Reviewed-by: David Ribeiro Alves <dralves@apache.org>
Reviewed-by: Mike Percy <mpercy@apache.org>