[java client] Extract RemoteTablet from AsyncKuduClient
RemoteTablet is responsible for handling the Java client's view of where replicas
are for its tablet, and who the leader is. Extracting this bit of functionality is
important if we want to be able to unit test it without bringing up a whole client
and possibly a cluster.

This patch makes a minor attempt at simplifying the interfaces, with more work to
come. Likewise for unit tests.

Change-Id: I3d06a573e4307c76a7aebab05cd5238fb0d9a2c6
Reviewed-on: http://gerrit.cloudera.org:8080/4719
Tested-by: Kudu Jenkins
Reviewed-by: Dan Burkert <danburkert@apache.org>