Enable GSSAPI for servers and ExternalMiniCluster
This adds a flag --server_require_kerberos, which makes the server only
advertise the GSSAPI mechanism. It also adds the appropriate hook to the
client to enable GSSAPI as a supported GSSAPI mechanism.

With this I was able to start a kudu master with the new flag enabled,
and then use Kerberos to authenticate a client.

This also adds some basic support to ExternalMiniCluster to start a
kerberized cluster. This involves starting a KDC, creating keytabs for
each server, passing the appropriate environment down as environment
variables, and setting up a client principal for the test itself.

In order for these tests to work, I had to tweak some settings in
krb5.conf, and then make corresponding changes to sasl_rpc-test to use
IP addresses instead of hostnames.

Change-Id: I595469e9cc8b2b2f57e9726014eeeb8112595801
Reviewed-on: http://gerrit.cloudera.org:8080/4765
Tested-by: Kudu Jenkins
Reviewed-by: Dan Burkert <danburkert@apache.org>
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>