KUDU-1680 - [python] Improve PartialRow Usability
The current semantics for setting values in a PartialRow are not
very elegant or comfortable for Python developers. This patch
improves this API by providing the ability to "load" a PartialRow
from a list, tuple or dictionary.  Some existing tests were modified,
however, some tests will continue to use the old API to ensure
backwards compatibility.

Change-Id: I5c9f57358e5048528398818fc80f32b27531a423
Reviewed-on: http://gerrit.cloudera.org:8080/4760
Tested-by: Kudu Jenkins
Reviewed-by: Will Berkeley <wdberkeley@gmail.com>