tool_action_test.cc: specify lambda return type
A user reported that they were unable to compile tool_action_test.cc on
OS X 10.11 with either the system Clang or thirdparty Clang. It appears
that these toolchains are not able to infer the correct return type for
a lambda. I'm not able to repro the failure, but the original user
confirmed that this patch fixed the issue. Failure:

    /Users/<user>/ProjectFiles/kudu/src/kudu/tools/tool_action_test.cc:384:5: error: return type 'kudu::Status' must match previous return type 'const ::kudu::Status' when lambda
          expression has unspecified explicit return type
        return Status::OK();
        ^
    1 error generated.

Change-Id: I52824e85a37c67bea8a2373264e8198d8d2294f4
Reviewed-on: http://gerrit.cloudera.org:8080/4773
Reviewed-by: Todd Lipcon <todd@apache.org>
Tested-by: Kudu Jenkins