KuduRPC integration with OpenSSL
This patch adds SSL support for the RPC layer in Kudu. It uses the
OpenSSL library for this purpose.

This is achieved by subclassing 'Socket' to add 'SSLSocket' which
calls into the OpenSSL library for auth/read/write/shutdown.

SSL is enabled only at a 'Connection' object level, i.e. the
AcceptorPool still only works with regular 'Socket's, and the reactor
threads also use a regular 'Socket' on a new outgoing call.

The first point of any SSL activity happens in the context of the
negotiation pool, where the SSL handshake happens before the SASL
handshake. On a successful handshake, further communication is
encrypted on that 'Connection'.

A 'SSLFactory' is created at the 'Messenger' level. This factory is
in-charge of creating 'SSLSocket' objects when necessary and also
keeps a track of the certificates, keys and the SSL context that is
shared among all the sockets of that 'Messenger'.

The x509_check_host() function is ported from OpenSSL-1.1.0b which we
use when the version of OpenSSL detected is less than 1.0.2. It is
added in util/x509_check_host.cc/h

Added a parameter to the rpc-test to additionally run the tests with
SSL enabled.

Future TODOs:
 - Allow loading keys as strings vs files. (Need to use different APIs)
 - Consider porting x509_check_ip and x509_check_ip_ascii.

Change-Id: I27167faa4e6a78e59b46093055b16682c93af0ea
Reviewed-on: http://gerrit.cloudera.org:8080/4789
Reviewed-by: Sailesh Mukil <sailesh@cloudera.com>
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>
Reviewed-by: Todd Lipcon <todd@apache.org>