KUDU-237 (part 1) - Support proper mutation encoding for reinserts
Currently reinserts are not properly encoded like the other mutations
and instead refer to an in-memory row, meaning we can't flush them
as they don't include the indirect data.

This patch makes it so that reinserts are properly encoded, like other
mutations but does not yet include the step of flushing them to disk,
which will be done in a follow-up patch.

Change-Id: I2b0a95e3f618bf20eb18ca4f4b03b9e2ce5704e8
Reviewed-on: http://gerrit.cloudera.org:8080/4791
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>