[java client] Remove timeouts when joining in the sync client
Folks have often seen the following exception:

CAUSED BY: TimeoutException: Timed out after 1ms when joining

followed by some gibberish about Deferreds. What happens here is that we short-circuit
the operation and we return to the user exactly when the timeout they provided expire.
This is nice because it means we respect their deadline to the millisecond. But it's
not nice because we don't know what's going on with their operation and in many cases
we have nothing to interrogate.

This patch removes the timeout when calling Deferred.join() so that the timeout instead
comes from the client internals with all the context we need. The two downsides are 1)
we will not be able to respect their timeout as closely and 2) if there's a bug in the
client then it might join forever. The latter means we expose bugs, so it's not that bad.

This patch also changes DeferredGroupException handling to make it more useful.

Change-Id: Iff4c0783e3c338d268ee1e37b000d45517b3880b
Reviewed-on: http://gerrit.cloudera.org:8080/4871
Reviewed-by: Adar Dembo <adar@cloudera.com>
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>