Remove unused code for checking PLAIN authentication
This removes the 'AuthStore' abstraction which theoretically supported
checking username/password authentication. In practice, we have no plans
to implement our own username/password authentication any time soon, so
this extra code just confuses the matter.

Additionally, this deprecates the user information that's passed as part
of the 'Connection Context'. The client still sends it, to satisfy older
servers, but the server ignores it and instead determines user
information from the SASL handshake.

Change-Id: Ie960fae30fe573b859f7ef0e27d656faac50d4c2
Reviewed-on: http://gerrit.cloudera.org:8080/4874
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>