Fix SASL mechanism negotiation on the server side
This fixes the order in which the SASL server is initialized such that
the mechanism list is set up before the sasl_server_init function is
called. This is equivalent to the change made in Change ID
I3c1b93045acd428ef3437597059c5106b03e25d0 on the client side.

Without this change, the server would use a null mechanism list, which
was being treated as a wildcard.

The new test case configures a server to only accept GSSAPI, and then
has a client configured to accept GSSAPI or PLAIN. Before the fix, it
would negotiate PLAIN despite the attempt to require GSSAPI on the
server side. Now, it properly negotiates.

Change-Id: I4741ce618525311258d5ba61dd85f6f4721caf66
Reviewed-on: http://gerrit.cloudera.org:8080/4875
Tested-by: Kudu Jenkins
Reviewed-by: Alexey Serbin <aserbin@cloudera.com>