[kudu-spark] Allow build against Spark 2.0 with Scala 2.11
This patch adds build profiles to build an artifact for Spark 2.0
with Scala 2.11. There was one tiny code change to package.scala
required to support Spark 2.0. To handle this, the different
versions of the package.scala reside in different directories and
the build profile includes the proper one.

The two artifactIds are
- kudu-spark_2.10
- kudu-spark2_2.11

The Jenkins job builds and tests both the default and spark2_2.11
profiles.

Change-Id: Ie055ac1e8f7be7cb992124afa0b601376db14f91
Reviewed-on: http://gerrit.cloudera.org:8080/4886
Tested-by: Kudu Jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@apache.org>