rpc: add a pre-flight check for server GSSAPI setup
This adds a pre-flight check that runs when adding an acceptor pool to a
messenger (which indicates that the messenger will act as a server). The
pre-flight check runs through a partial SASL GSSAPI authentication flow
in order to verify that the plugin is set up and a keytab is available.

Tested this manually as well as with a new unit test.

Change-Id: Ia26482003144504dee77ffb2e9de7ebe3c1e2e3d
Reviewed-on: http://gerrit.cloudera.org:8080/4909
Tested-by: Kudu Jenkins
Reviewed-by: Todd Lipcon <todd@apache.org>