Make RowChangeListDecoder::RemoveColumnIdsFromChangeList() take a RowChangeListEncoder as an out param
All the other methods in this class, of this genre, take one but
this one was taking the raw faststring buffer instead. Though it
did have the advantage of not forcing the caller to clear() the
buffer everytime it was weirdly out of place and might introduce
subtle bugs with new semantics for RowChangeListEncoder.

Change-Id: I6584b22662b58227a80eaebb7d89199d3ea2498a
Reviewed-on: http://gerrit.cloudera.org:8080/4928
Reviewed-by: Todd Lipcon <todd@apache.org>
Tested-by: David Ribeiro Alves <dralves@apache.org>