KUDU-1677. Scanners can return out-of-order rows during a compaction
This fixes a bug where a scanner could return out-of-order rows if
started during the "duplicating" phase of a compaction.

The issue was that DuplicatingRowSet would create a UnionIterator rather
than a MergeIterator despite the fact that the root iterator was
supposed to be in an ordered mode.

Fixing this involved passing the OrderMode down into the NewRowIterator
call. This argument is only used by the DuplicatingRowSet for now, but
may be useful for other rowset types down the road (e.g. the MemRowSet
might be able to scan faster if it could scan in insertion order instead
of in key order).

I also removed the redundant Tablet::OrderMode enum and instead just
passed the protobuf-based enum (from common.proto) all the way through,
which simplified things a little bit.

Change-Id: I6d5c373e2d56039a3b8be272cb4fea3d5d840e5b
Reviewed-on: http://gerrit.cloudera.org:8080/4983
Tested-by: Kudu Jenkins
Reviewed-by: Adar Dembo <adar@cloudera.com>