Account for REINSERTs in delta stats
This makes DeltaStats also account for REINSERTs. This is not
actually used outside of tests, but seems like it would be
silly to not count this type of delta. In the future it might
be useful for selecting undo deltas for minor delta compaction.

The protobuf field is optional to preserve data format
compatibility.

Change-Id: Idd60f6c1c12803d339f5f8d96c6b089fab21b13f
Reviewed-on: http://gerrit.cloudera.org:8080/4932
Tested-by: Kudu Jenkins
Reviewed-by: Jean-Daniel Cryans <jdcryans@apache.org>